ADD excepciones Odontologo y turno

+++ src/main/java/org/e7/clinica/service/impl/OdontologoService.java
@@ -1,6 +1,8 @@
 package org.e7.clinica.service.impl;
 
 import org.e7.clinica.entity.Odontologo;
+import org.e7.clinica.entity.Paciente;
+import org.e7.clinica.exception.ResourceNotFoundException;
 import org.e7.clinica.service.IOdontologoService;
 import org.springframework.stereotype.Service;
 import org.e7.clinica.repository.IOdontologoRepository;
@@ -24,22 +26,43 @@ public class OdontologoService implements IOdontologoService {
 
     @Override
     public Optional<Odontologo> buscarPorId(Integer id) {
-        return iOdontologoRepository.findById(id);
+        Optional<Odontologo> odontologo = iOdontologoRepository.findById(id);
+        if (odontologo.isPresent()) {
+            return odontologo;//Repository.findById(id);
+        } else {
+            throw new ResourceNotFoundException("El odontologo con ID  " + id + "  no fue encontrado");
+        }
     }
 
     @Override
     public List<Odontologo> buscarTodos() {
+        List<Odontologo>    odontologos=iOdontologoRepository.findAll();
+        if(odontologos.isEmpty()){
+            throw new ResourceNotFoundException("No se encontraron odontologos");
+        }
         return iOdontologoRepository.findAll();
     }
 
     @Override
     public void modificarOdontologo(Odontologo odontologo) {
-        iOdontologoRepository.save(odontologo);
+        Optional<Odontologo> odontologoEncontrado = iOdontologoRepository.findById(odontologo.getId());
+        if(odontologoEncontrado.isPresent()){
+            System.out.println("Odontólogo encontrado, actualizando: " + odontologo.getId());
+            iOdontologoRepository.save(odontologo);
+        } else {
+            System.out.println("Odontólogo no encontrado con ID: " + odontologo.getId());
+            throw new ResourceNotFoundException("El odontologo  no fue encontrado");
+        }
     }
 
     @Override
     public void eliminarOdontologo(Integer id) {
-        iOdontologoRepository.deleteById(id);
+        Optional<Odontologo> odontologoEncontrado = iOdontologoRepository.findById(id);
+        if(odontologoEncontrado.isPresent()){
+            iOdontologoRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("El Odontologo  con ID "+ id +" no fue encontrado");
+        }
     }
 
     @Override
@@ -56,7 +79,6 @@ public class OdontologoService implements IOdontologoService {
     public Long contarOdontologos() {
         return iOdontologoRepository.contarOdontologos();
     }
-
 
 }
